Split browseContexts, postCases, Filterline, Filter, ExternalFilter from filteReddit.js to separate files
