+++ lib/modules/filteReddit/ExternalFilter.js
@@ -1,0 +1,33 @@
+import _ from 'lodash';
+import { $ } from '../../vendor';
+import * as Options from '../../core/options';
+import * as FilteReddit from '../filteReddit';
+import filterlineExternalFilterTemplate from '../../templates/filterlineExternalFilter.mustache';
+import Filter from './Filter';
+
+export default class ExternalFilter extends Filter {
+	createElement() {
+		this.element = $(filterlineExternalFilterTemplate(this))[0];
+	}
+
+	removeFilterEntry(entry) {
+		const newValueArray = _.pull(FilteReddit.module.options[this.key].value, entry);
+		Options.set(FilteReddit.module, this.key, newValueArray);
+		if (this.clearCache) this.clearCache();
+		this.parent.updateFilter(this);
+	}
+
+	setThingFilterReason(thing) {
+		thing.element.setAttribute('filter-reason', `specified by filter ${this.key}`);
+		thing.filterEntryRemover = $('<span>', {
+			class: 'res-filter-remove-entry',
+			title: JSON.stringify(thing.filterResult, null, '  '),
+			click: () => this.removeFilterEntry(thing.filterResult),
+		}).prependTo(thing.element);
+	}
+
+	removeThingFilterReason(thing) {
+		thing.element.removeAttribute('filter-reason');
+		if (thing.filterEntryRemover) thing.filterEntryRemover.remove();
+	}
+}
